Object::spawnScene() - 100%

+++ src/Bases/Base.cpp
@@ -367,7 +367,7 @@ Base *Base::spawnChild(u16 overlay_id, Base *b, u32 c, u8 d)
   }
   return (Base *)0x00;
 }
-Base *Base::spawnParent(u16 a, u32 b, u8 c)
+Base *Base::spawnParent(u32 a, u32 b, u32 c)
 {
   return Base::spawn(a, 0, b, c);
 }
